+++ src/webapplication.cpp
@@ -5,8 +5,8 @@
 #include "httprequest.h"
 #include "http.h"
 #include "thread.h"
+#include "utils.h"
 #include <Rinternals.h>
-
 
 std::string normalizeHeaderName(const std::string& name) {
   std::string result = name;
@@ -128,11 +128,21 @@ void requestToEnv(boost::shared_ptr<HttpRequest> pRequest, Rcpp::Environment* pE
   env["REMOTE_PORT"] = CharacterVector(rportstr.str());
 
   const RequestHeaders& headers = pRequest->headers();
+  Rcpp::CharacterVector raw_headers(headers.size());
+  Rcpp::CharacterVector raw_header_names(headers.size());
+
   for (RequestHeaders::const_iterator it = headers.begin();
     it != headers.end();
     it++) {
+    int idx = std::distance(headers.begin(), it);
     env["HTTP_" + normalizeHeaderName(it->first)] = CharacterVector(it->second);
+    raw_header_names[idx] = to_lower(it->first);
+    raw_headers[idx] = it->second;
   }
+  raw_headers.attr("names") = raw_header_names;
+
+  env["HEADERS"] = raw_headers;
+
 }